feat: add InitializeEquilibriumStates component and refactor infection seeding process
